+++ src/cli/command-registrar.ts
@@ -119,7 +119,7 @@ function execAction(command, concrete, args) {
 }
 
 function serializeErrAndExit(err, commandName) {
-  process.stderr.write(packCommand(buildCommandMessage(serializeError(err)), false));
+  process.stderr.write(packCommand(buildCommandMessage(serializeError(err), undefined, false), false, false));
   const code = err.code && isNumeric(err.code) ? err.code : 1;
   return logger.exitAfterFlush(code, commandName);
 }
@@ -237,9 +237,9 @@ export default class CommandRegistrar {
     const aliasList = this.commands.map(cmd => first(cmd.alias.split(' ')));
 
     if (
-      cmdList.includes(subcommand) &&
-      extensionsCmdList.includes(subcommand) &&
-      aliasList.includes(subcommand) &&
+      !cmdList.includes(subcommand) &&
+      !extensionsCmdList.includes(subcommand) &&
+      !aliasList.includes(subcommand) &&
       subcommand !== '-V' &&
       subcommand !== '--version'
     ) {