+++ src/constants.ts
@@ -224,6 +224,8 @@ export const CFG_COMMAND_INTERACTIVE_TEMPLATE = 'interactive.{commandName}';
 
 export const CFG_INIT_INTERACTIVE = format(CFG_COMMAND_INTERACTIVE_TEMPLATE, { commandName: INIT_COMMAND });
 
+export const CFG_SSH_NO_COMPRESS = 'ssh_no_compress';
+
 /**
  * git hooks
  */
@@ -401,3 +403,7 @@ export const WILDCARD_HELP = (command: string) =>
 export const CURRENT_UPSTREAM = 'current';
 
 export const DEPENDENCIES_FIELDS = ['dependencies', 'devDependencies', 'peerDependencies'];
+
+const MISSING_DEPS_SPACE_COUNT = 10;
+export const MISSING_DEPS_SPACE = ' '.repeat(MISSING_DEPS_SPACE_COUNT);
+export const MISSING_NESTED_DEPS_SPACE = ' '.repeat(MISSING_DEPS_SPACE_COUNT + 2);